New autonomous subroutine
autoRed1() now makes the robot follow the key line into the hopper.
autoRed1A() is the old drive forward, turn, drive forward to the hopper
routine.
the timedDrive() command should allow the robot to drive with different
speeds on the right and left sides for a certain amount of time.