Auto Calibration (guesses) for State Champs
Based on testing done at Bulldogs I made the guesstimate that the blue
side gear routine would be the same for the red side except for reversed
turns. And the side gear autos would also just be the same numbers as
the first parts of the blue side gear routine.